feat: core refactor
Instead of having a giga-sized manager.go file with 2000+ lines, I've broken it down into smaller modules grouped by function. This will make the maintenance easier, way easier.
