+++ internal/core/services/post_processor.go
@@ -1,4 +1,4 @@
-package core
+package services
 
 import (
 	"context"
