+++ internal/core/services/torrent_selector.go
@@ -1,4 +1,4 @@
-package core
+package services
 
 import (
 	"fmt"
